pyproject.toml
[project]
name = "ai-calc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.0",
    "langchain-openai>=1.1.4",
    "langgraph>=1.0.5",
    "python-dotenv>=1.2.1",
]
